Update dockerfile to use OpenJDK from Docker Hub
We moved the majority of images away from GHCR to DockerHub a while ago
it seems that test-rp wasn't shown the same love.  So this PR fixes that
issue and also allows us to inject a docker image name using build args
which is useful for testing and for verify-local-startup.